[2020/06/09] queryDSL, 사용예제 추가.

+++ src/main/java/com/simple/portal/biz/v1/board/api/BoardMemAPI.java
@@ -32,7 +32,7 @@ public class BoardMemAPI {
      * @return
      */
     @GetMapping("/board/{seq}")
-    public ResponseEntity<ApiResponse> detail(@PathVariable int seq) {
+    public ResponseEntity<ApiResponse> detail(@PathVariable Long seq) {
 
         BoardEntity boardEntity = new BoardEntity();
         boardEntity.setId(seq);
